backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "IBE160 Backend"
authors = ["BIP <bip@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.122.0"
uvicorn = "^0.38.0"
supabase = "^2.24.0"
python-dotenv = "^1.2.1"
email-validator = "^2.1.1"
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.44"
sqlmodel = "^0.0.27"
psycopg2-binary = "^2.9.11"
asyncpg = "^0.31.0"
pypdf = "^6.4.0"
python-docx = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.1,<10.0.0"
httpx = ">=0.28.1,<0.29.0"
reportlab = {version = "4.4.5", python = ">=3.10,<4"}
